// src/app/model/healthExamination.ts
export class HealthExamination {
  dateOfExamination: Date;
  diagnosisAtMKB: string;
  laboratory_Finding: boolean;
  RTC_Finding: boolean;
  EHO_Finding: boolean;
  computed_Tomography: boolean;
  magnetic_Resonance: boolean;
  sent_To_A_specialist: boolean;
  illness: boolean;
  bandage: boolean;
  antibody_Prophylaxis: boolean;
  typeOfTherapy: string;


  constructor() {

  }
}